requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
pika==1.3.2
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.7
